+++ src/ansibleguy-webui/aw/model/base.py
@@ -6,15 +6,17 @@ CHOICES_BOOL = (
 )
 DEFAULT_NONE = {'null': True, 'default': None, 'blank': True}
 JOB_EXEC_STATUS_SUCCESS = 4
+JOB_EXEC_STATUS_FAILED = 3
 CHOICES_JOB_EXEC_STATUS = [
     (0, 'Waiting'),
     (1, 'Starting'),
     (2, 'Running'),
-    (3, 'Failed'),
+    (JOB_EXEC_STATUS_FAILED, 'Failed'),
     (JOB_EXEC_STATUS_SUCCESS, 'Finished'),
     (5, 'Stopping'),
     (6, 'Stopped'),
 ]
+JOB_EXEC_STATUS_ACTIVE = [0, 1, 2, 5]
 
 
 class BareModel(models.Model):